requirements.txt
Django==3.2.20
djangorestframework==3.14.0
django-cors-headers==4.2.0
drf-yasg==1.21.7
django-filter==23.2
mysqlclient==2.2.0
redis==4.6.0
django-redis==5.3.0
celery==5.2.7
django_celery_beat==2.5.0
django_celery_results==2.5.1
dj-database-url==2.0.0
sentry-sdk==1.29.2
gunicorn==21.2.0
python-dotenv==1.0.0
django-debug-toolbar==4.1.0
requests==2.31.0
mongoengine==0.27.0
python-json-logger==2.0.7
djangorestframework-api-key==2.0.0
cryptography==41.0.7